PRO-2762: Add functional tests for the submit controller 	  - Implement functional tests for submit and resubmit methods 	  - Add method to handle submit headers (this was added to TestUtils)
